backend/requirements.txt
fastapi>=0.109.0,<0.110.0
uvicorn>=0.27.0,<0.28.0
motor>=3.3.0,<4.0.0
redis>=4.2.0,<5.0.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
python-multipart>=0.0.7,<0.1.0
aiofiles>=23.2.0,<24.0.0
httpx>=0.26.0,<0.27.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
click>=8.0.0
rich>=13.0.0
requests>=2.31.0
asyncpg>=0.29.0

# Monitoramento e Métricas
prometheus-client>=0.19.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.43b0

# Segurança
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Cache e Rate Limiting
fastapi-limiter>=0.1.5
cachetools>=5.3.2

# Desenvolvimento e Testes
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
black>=24.1.0
flake8>=7.0.0
mypy>=1.8.0
isort>=5.13.0
pytest-benchmark>=4.0.0
locust>=2.24.0
faker>=22.0.0
